requirements.txt
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
langgraph==0.2.60
faiss-cpu==1.9.0.post1
chromadb==0.6.3
sentence-transformers==3.3.1
fastapi==0.115.6
uvicorn==0.32.1
streamlit==1.41.1
python-dotenv==1.0.1
pydantic==2.10.4
